+++ arctic/chunkstore/chunkstore.py
@@ -24,7 +24,6 @@ APPEND_COUNT = 'ac'
 ROWS = 'r'
 
 
-
 class ChunkStore(object):
     @classmethod
     def initialize_library(cls, arctic_lib, **kwargs):
@@ -73,12 +72,11 @@ class ChunkStore(object):
     def __repr__(self):
         return str(self)
 
-    def _checksum(self, symbol, doc):
+    def _checksum(self, doc):
         """
         Checksum the passed in dictionary
         """
         sha = hashlib.sha1()
-        sha.update(symbol.encode('ascii'))
         sha.update(self.chunker.chunk_to_str(doc[START]).encode('ascii'))
         sha.update(self.chunker.chunk_to_str(doc[END]).encode('ascii'))
         for k in doc[DATA][COLUMNS]:
@@ -126,6 +124,31 @@ class ChunkStore(object):
     def _get_symbol_info(self, symbol):
         return self._symbols.find_one({SYMBOL: symbol})
 
+    def rename(self, from_symbol, to_symbol):
+        """
+        Rename a symbol
+
+        Parameters
+        ----------
+        from_symbol: str
+            the existing symbol that will be renamed
+        to_symbol: str
+            the new symbol name
+        """
+
+        sym = self._get_symbol_info(from_symbol)
+        if not sym:
+            raise NoDataFoundException('No data found for %s' % (from_symbol))
+
+        if self._get_symbol_info(to_symbol) is not None:
+            raise Exception('Symbol %s already exists' % (to_symbol))
+
+        mongo_retry(self._collection.update_many)({SYMBOL: from_symbol},
+                                                  {'$set': {SYMBOL: to_symbol}})
+
+        mongo_retry(self._symbols.update_one)({SYMBOL: from_symbol},
+                                              {'$set': {SYMBOL: to_symbol}})
+
     def read(self, symbol, chunk_range=None, columns=None, filter_data=True):
         """
         Reads data for a given symbol from the database.
@@ -160,7 +183,7 @@ class ChunkStore(object):
             spec.update(self.chunker.to_mongo(chunk_range))
 
         segments = []
-        for _, x in enumerate(self._collection.find(spec, sort=[(START, pymongo.ASCENDING)],)):
+        for x in self._collection.find(spec, sort=[(START, pymongo.ASCENDING)],):
             segments.append(x[DATA])
 
         data = self.serializer.deserialize(segments, columns)
@@ -192,7 +215,7 @@ class ChunkStore(object):
         doc[CHUNK_SIZE] = chunk_size
         doc[ROWS] = len(item)
         doc[TYPE] = 'dataframe' if isinstance(item, DataFrame) else 'series'
-        
+
         sym = self._get_symbol_info(symbol)
         if sym:
             previous_shas = set([Binary(x[SHA]) for x in self._collection.find({SYMBOL: symbol},
@@ -212,7 +235,7 @@ class ChunkStore(object):
             chunk[START] = start
             chunk[END] = end
             chunk[SYMBOL] = symbol
-            chunk[SHA] = self._checksum(symbol, chunk)
+            chunk[SHA] = self._checksum(chunk)
 
             if chunk[SHA] not in previous_shas:
                 op = True
@@ -248,12 +271,11 @@ class ChunkStore(object):
         sym = self._get_symbol_info(symbol)
         if not sym:
             raise NoDataFoundException("Symbol does not exist.")
-        
+
         if sym[TYPE] == 'series' and not isinstance(item, Series):
             raise Exception("Cannot combine Series and DataFrame")
         if sym[TYPE] == 'dataframe' and not isinstance(item, DataFrame):
             raise Exception("Cannot combine DataFrame and Series")
-
 
         bulk = self._collection.initialize_unordered_bulk_op()
         op = False
@@ -278,19 +300,19 @@ class ChunkStore(object):
             data = self.serializer.serialize(record)
             op = True
 
-            segment = {DATA: data}
-            segment[TYPE] = 'dataframe' if isinstance(record, DataFrame) else 'series'
-            segment[START] = start
-            segment[END] = end
-            sha = self._checksum(symbol, segment)
-            segment[SHA] = sha
+            chunk = {DATA: data}
+            chunk[TYPE] = 'dataframe' if isinstance(record, DataFrame) else 'series'
+            chunk[START] = start
+            chunk[END] = end
+            sha = self._checksum(chunk)
+            chunk[SHA] = sha
             if new_chunk:
                 # new chunk
                 bulk.find({SYMBOL: symbol, SHA: sha}
-                          ).upsert().update_one({'$set': segment})
+                          ).upsert().update_one({'$set': chunk})
             else:
                 bulk.find({SYMBOL: symbol, START: start, END: end}
-                          ).update_one({'$set': segment})
+                          ).update_one({'$set': chunk})
 
         if op:
             bulk.execute()